+++ src/xrayradar_server/main.py
@@ -77,7 +77,7 @@ async def lifespan(_: FastAPI):
 
 app = FastAPI(
     title="xrayradar-server",
-    version="0.13.0",
+    version="0.14.0",
     lifespan=lifespan,
     docs_url=None,  # Disable default docs
     redoc_url=None,  # Disable default redoc